# _doc/swagger.yaml
openapi: 3.0.0
servers:
    -   url: '{protocol}://{api_host}/sales-channel-api/v1'
        variables:
            protocol:
                default: 'https'
                description: transfer protocol (HTTP or HTTPS)
            api_host:
                default: 'shopware-platform.test'
                description: 'Host of your Shopware 6 instance'
info:
    title: Shopware Sales-Channel-API Extension
    version: 1.0.0
paths:
...
    '/vsf/navigation':
        post:
            summary: Get navigation tree for a given root node
            operationId: vsf.navigation
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                depth:
                                    type: integer
                                    description: 'Depth of the tree (1 = only  fetch first level children; -1 = fetch full tree)'
                                rootNode:
                                    type: string
                                    format: uuid
                                    description: 'Relative root for the tree to be fetched from'
            responses:
                '200':
                    description: Resolved navigation tree
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    count:
                                        type: integer
                                    elements:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/navigation_tree_item'
components:
    schemas:
        navigation_tree_item:
            type: object
            properties:
                route:
                    type: object
                    properties:
                        path:
                            type: string
                        resourceType:
                            type: string
                name:
                    type: string
                children:
                    type: array
                    items:
                        $ref: '#/components/schemas/navigation_tree_item'
                count:
                    type: integer
                level:
                    type: integer
...
    securitySchemes:
        ApiKey:
            type: apiKey
            in: header
            name: sw-access-key
security:
    - ApiKey: []